Stand on the shoulders of giants for -fvisibility=hidden
I have discovered a macro that was furthermore already installed in m4/,
that allows us to remove our own logic and benefit from more checking.
The file visibility.m4 is inexplicably provided by gettext-devel in
Fedora, but as we already need that for builds, I will not add the m4
file to the tracked sources.